Move SwitchRow into components
