+++ ts/components/ConfigInput.svelte
