+++ ts/components/Label.svelte
