+++ ts/deck-options/icons.ts
@@ -4,5 +4,4 @@
 /// <reference types="../lib/image-import" />
 
 export { default as chevronDown } from "@mdi/svg/svg/chevron-down.svg";
-export { default as revertIcon } from "bootstrap-icons/icons/arrow-counterclockwise.svg";
 export { default as gearIcon } from "bootstrap-icons/icons/gear.svg";
